Move reducers to state/reducers and related changes
